pyproject.toml
[project]
name = "onlinebooking"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.8.1",
    "flake8>=7.3.0",
    "orjson>=3.11.3",
    "redis>=6.4.0",
]
